Improved HSC logic.
Added manual translation multiplier that depends on the angle between MaxMT and
current Fwd vector. This reduces lateral velocity gain when a vessel starts to a
target to which it has to turn cosiderably (> 90deg).
Also HSC does not perform thrust vector rotation when starting to a target while
having manual thrust. To put it simply, if you're idle, first turn to your
target, then start to move.