Changed the way torqueRatio is calculated
Now it is a measure of relation between the amount of angular acceleration and
linear acceleration towards CoM that an engine produces.

It works as follows:
* from two identical engines the one that produces more torque will be preferred
during optimization.
* from two different engines, one with weak thrust but large torque, the other
with strong thrust and the same torque, the former will be preferred.

This allows building maneuver-only engine systems that have weak attitude
thrusters with large levers and strong translational CoM-aligned thrusters.