EngineWrapper added isThruster and thrustLimiterLocked flags
Rationale:

isVSC flag indicates that the engine is used by VSC to control anti-gravity
thrust

But MANUAL engines (and thus engines with throttleLocked=true, e.g. SRBs, for
which the Role is forced) cannot be used by VSC, yet they still generate thrust
that should be accounted for (e.g. for takeoff/landing and maneuvers). Such
engines are marked (along with all VSC capable engines) with isThruster.

On the other hand, there's a difference between ModuleEngine.throttleLocked and
MANUAL mode: the former locks mainThottle for that particular engine at 100%,
while thrust limiter of a MANUAL can be changed by user, but not by TCA itself.
So we need to mark MANUAL engines with thrustLimiterLocked flag to correctly
calculate various derived values.